+++ index.html
@@ -235,6 +235,6 @@
     </template>
 
   <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
+  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
