+++ js/fake-data/const.js
@@ -1,0 +1,73 @@
+const MIN_COMMENTS_COUNT = 0;
+const MAX_COMMENTS_COUNT = 30;
+const MIN_LIKES_COUNT = 15;
+const MAX_LIKES_COUNT = 200;
+const MIN_AVATAR_INDEX = 1;
+const MAX_AVATAR_INDEX = 6;
+const MIN_SENTENCES_IN_COMMENT = 1;
+const MAX_SENTENCES_IN_COMMENT = 2;
+
+
+const PHOTO_DESCRIPTIONS = {
+  1: 'Санаторий',
+  2: 'GOTOTHEBEACH',
+  3: 'Пляж',
+  4: 'Жоньщина',
+  5: 'Весёлый суп',
+  6: 'Машина (черная)',
+  7: 'Ужин на двоих',
+  8: 'Киселик',
+  9: 'Анапа',
+  10: 'Тапки',
+  11: 'Какой-то забор',
+  12: 'Фото автомобиля',
+  13: 'Нечто',
+  14: 'Ролл с котом',
+  15: 'Бархатные тяги',
+  16: 'Самолет',
+  17: 'ААААААААА!',
+  18: 'Машина (красная)',
+  19: 'Тапки дайсон',
+  20: 'Деревья',
+  21: '1 миска рис',
+  22: 'Закат',
+  23: 'Краб',
+  24: 'Руки вверх',
+  25: 'Бегемот орет'
+};
+
+const USER_NAMES = [
+  'Артем',
+  'Никита',
+  'Даша',
+  'Полина',
+  'Аркадий',
+  'Вика',
+  'Аня',
+  'Таня',
+  'Саша',
+  'Маша',
+];
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {
+  MIN_COMMENTS_COUNT,
+  MAX_COMMENTS_COUNT,
+  MIN_LIKES_COUNT,
+  MAX_LIKES_COUNT,
+  MIN_AVATAR_INDEX,
+  MAX_AVATAR_INDEX,
+  MIN_SENTENCES_IN_COMMENT,
+  MAX_SENTENCES_IN_COMMENT,
+  PHOTO_DESCRIPTIONS,
+  USER_NAMES,
+  COMMENT_MESSAGES
+};
